Move slice input class
